requirements.txt
aws-cdk-lib==2.177.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0,<1.27.0
